improved validation

+++ backend/controllers/reviews.js
@@ -1,6 +1,12 @@
 const mongodb = require('../database/connect');
 const ObjectId = require('mongodb').ObjectId;
 
+/*********
+ * Regex *
+ *********/
+var rgx_letters = /^[A-Za-z\s]+$/;
+var rgx_email = /^\S+@\S+\.\S+$/;
+
 /***********************
  * To get all reviews: *
  *    Takes no params  *
@@ -28,7 +34,7 @@ const getAll = async (req, res) => {
  *****************************/
 const getReviewID = (req, res) => {
   if (!ObjectId.isValid(req.params.id)) {
-    res.status(400).json('Must pass a valid ID to delete a menu item.');
+    res.status(400).json('Must pass a valid ID to get a review.');
   } else {
     const reviewId = new ObjectId(req.params.id);
     mongodb
@@ -56,11 +62,19 @@ const getReviewID = (req, res) => {
 const getReviewEmail = async (req, res) => {
   try {
     const email = req.params.email.toLowerCase();
-    const response = await mongodb.getDb().db().collection('reviews').find({ email: email });
-    response.toArray().then((lists) => {
-      res.setHeader('Content-Type', 'application/json');
-      res.status(200).json(lists);
-    });
+    if (!email.match(rgx_email)) {
+      res.status(412).json(`The email format is invalid.`);
+    } else {
+      const response = await mongodb.getDb().db().collection('reviews').find({ email: email });
+      response.toArray().then((lists) => {
+        if (lists.length === 0) {
+          res.status(404).json(`No reviews were found with email: ${email}`);
+        } else {
+          res.setHeader('Content-Type', 'application/json');
+          res.status(200).json(lists);
+        }
+      });
+    }
   } catch (err) {
     res.status(500).json(err);
   }
@@ -76,8 +90,12 @@ const getReviewState = async (req, res) => {
     const state = req.query.state.toLowerCase();
     const response = await mongodb.getDb().db().collection('reviews').find({ state: state });
     response.toArray().then((lists) => {
-      res.setHeader('Content-Type', 'application/json');
-      res.status(200).json(lists);
+      if (lists.length === 0) {
+        res.status(404).json(`No reviews were found on state: ${state}`);
+      } else {
+        res.setHeader('Content-Type', 'application/json');
+        res.status(200).json(lists);
+      }
     });
   } catch (err) {
     res.status(500).json(err);
@@ -122,7 +140,7 @@ const createReview = async (req, res) => {
  ****************************/
 const editReviewID = async (req, res) => {
   if (!ObjectId.isValid(req.params.id)) {
-    res.status(400).json('Must pass a valid ID to delete a menu item.');
+    res.status(400).json('Must pass a valid ID to update a menu item.');
   } else {
     const reviewId = new ObjectId(req.params.id);
     const newReview = {
@@ -165,31 +183,37 @@ const editReview = async (req, res) => {
     const state = req.params.state.toLowerCase();
     // Lowercase email, city, state because they are used on database
     // querrying, case sensitive search
-    const newReview = {
-      firstName: req.body.firstName,
-      lastName: req.body.lastName,
-      phone: req.body.phone,
-      email: req.params.email.toLowerCase(),
-      review: req.body.review,
-      stars: req.body.stars,
-      city: req.params.city.toLowerCase(),
-      state: req.params.state.toLowerCase()
-    };
-    const response = await mongodb
-      .getDb()
-      .db()
-      .collection('reviews')
-      .replaceOne({ email: email, city: city, state: state }, newReview);
+    // validate user's input
+    if (!email.match(rgx_email)) res.status(412).json(`The email format is invalid.`);
+    else if (!city.match(rgx_letters)) res.status(412).json(`The city format is invalid.`);
+    else if (!state.match(rgx_letters)) res.status(412).json(`The state format is invalid.`);
+    else {
+      const newReview = {
+        firstName: req.body.firstName,
+        lastName: req.body.lastName,
+        phone: req.body.phone,
+        email: req.params.email.toLowerCase(),
+        review: req.body.review,
+        stars: req.body.stars,
+        city: req.params.city.toLowerCase(),
+        state: req.params.state.toLowerCase()
+      };
+      const response = await mongodb
+        .getDb()
+        .db()
+        .collection('reviews')
+        .replaceOne({ email: email, city: city, state: state }, newReview);
 
-    if (response.modifiedCount > 0) {
-      res.status(204).send();
-    } else {
-      res
-        .status(500)
-        .json(
-          response.error ||
-            `Something went wrong while attempting to update review with email ${email}, city ${city} and state ${state}`
-        );
+      if (response.modifiedCount > 0) {
+        res.status(204).send();
+      } else {
+        res
+          .status(500)
+          .json(
+            response.error ||
+              `Something went wrong while attempting to update review with email ${email}, city ${city} and state ${state}`
+          );
+      }
     }
   } catch (err) {
     res.status(500).json(err);
@@ -232,21 +256,27 @@ const deleteReview = async (req, res) => {
     const email = req.params.email.toLowerCase();
     const city = req.params.city.toLowerCase();
     const state = req.params.state.toLowerCase();
-    const response = await mongodb
-      .getDb()
-      .db()
-      .collection('reviews')
-      .deleteOne({ email: email, city: city, state: state }, true);
+    // validate user's input
+    if (!email.match(rgx_email)) res.status(412).json(`The email format is invalid.`);
+    else if (!city.match(rgx_letters)) res.status(412).json(`The city format is invalid.`);
+    else if (!state.match(rgx_letters)) res.status(412).json(`The state format is invalid.`);
+    else {
+      const response = await mongodb
+        .getDb()
+        .db()
+        .collection('reviews')
+        .deleteOne({ email: email, city: city, state: state }, true);
 
-    if (response.deletedCount > 0) {
-      res.status(204).send();
-    } else {
-      res
-        .status(500)
-        .json(
-          response.error ||
-            `Something went wrong while attempting to delete review with email ${email}, city ${city} and state ${state}`
-        );
+      if (response.deletedCount > 0) {
+        res.status(204).send();
+      } else {
+        res
+          .status(500)
+          .json(
+            response.error ||
+              `Something went wrong while attempting to delete review with email ${email}, city ${city} and state ${state}`
+          );
+      }
     }
   } catch (err) {
     res.status(500).json(err);
